ContactsPicker refactoring logic into composable
